First attempt at making a website
Not even for this project, but in general.
This is literally my first attempt at that.  Ever.
Thanks Bootstrap for making this look like non-crap.
I'm fully aware that the repetitions are dumb, but I don't want to do that by JS now,
mostly because I have not the faintest idea how to do that.
Also, yeah, there should be images, each thingy should be centered, etc.
I know, it's crap.
Again, thanks to Bootstrap it doesn't suck that much.